+++ pyttern/language_processors/language_utils.py
@@ -1,0 +1,11 @@
+def determine_language(filename):
+    """
+    Determines the language based on the file extension.
+    Returns 'python' or 'java' atm or None for unsupported file types.
+    """
+    extension = filename.split('.')[-1]
+    if extension in {"pyt", "py"}:
+        return "python"
+    elif extension in {"jat", "java"}:
+        return "java"
+    return None
